All issues raised in comments are resolved
